add list fixture specs
