+++ examples/fundamentals__fixtures/cypress/fixtures/users.json
@@ -1,0 +1,14 @@
+[
+  {
+    "name": "Joe Bravo",
+    "age": 20
+  },
+  {
+    "name": "Mike Smith",
+    "age": 30
+  },
+  {
+    "name": "Alicia Blue",
+    "age": 25
+  }
+]
